Modification for the pipeline

+++ sirh-backend/src/main/java/com/example/anywrpfe/services/implementations/TokenServiceImpl.java
@@ -5,11 +5,10 @@ import com.example.anywrpfe.entities.Token;
 import com.example.anywrpfe.repositories.CollaborateurRepository;
 import com.example.anywrpfe.repositories.TokenRepository;
 import com.example.anywrpfe.services.TokenService;
-import lombok.AllArgsConstructor;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.Calendar;
 import java.util.Optional;
 
 @Service
@@ -20,25 +19,25 @@ public class TokenServiceImpl implements TokenService {
     private final CollaborateurRepository collaborateurRepository;
 
 
-
     @Override
     public String tokenValidate(String token) {
-        Optional<Token> OptionalToken = tokenRepository.findByToken(token);
+        Optional<Token> optionalToken = tokenRepository.findByToken(token);
 
-
-
-            Token theToken = OptionalToken.get();
-            Collaborateur collaborateur = theToken.getCollaborateur();
-            if(theToken.expired)
-            {
-                tokenRepository.delete(theToken);
-                return "Token expired";
-            }
-
-            collaborateurRepository.save(collaborateur);
-
-            return "valid";
+        if (optionalToken.isEmpty()) {
+            throw new EntityNotFoundException("Token not found or invalid: " + token);
         }
+
+        Token theToken = optionalToken.get();
+        Collaborateur collaborateur = theToken.getCollaborateur();
+
+        if (theToken.expired) {
+            tokenRepository.delete(theToken);
+            return "Token expired";
+        }
+
+        collaborateurRepository.save(collaborateur);
+        return "valid";
     }
+}
